+++ Joobie/Joobie/Data/Repositories/Interfaces/IRepository.cs
@@ -7,7 +7,7 @@ namespace Shop.Data.Repositories
 {
     public interface IRepository<TEntity> where TEntity : class
     {
-        Task<TEntity> GetAsync(int id);
+        Task<TEntity> GetAsync(long id);
         Task<IEnumerable<TEntity>> GetAllAsync();
         Task<IEnumerable<TEntity>> FindAsync(Expression<Func<TEntity, bool>> predicate);
         Task<TEntity> SingleOrDefaultAsync(Expression<Func<TEntity, bool>> predicate);
